+++ packages/lit-helpers/src/spreadProps.d.ts
@@ -1,0 +1,3 @@
+import { PropertyPart } from 'lit-html';
+
+export declare const spreadProps: (props: { [key: string]: unknown }) => (part: PropertyPart) => void;