+++ packages/rollup-plugin-index-html/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.9.3](https://github.com/open-wc/open-wc/compare/rollup-plugin-index-html@1.9.2...rollup-plugin-index-html@1.9.3) (2020-02-23)
+
+**Note:** Version bump only for package rollup-plugin-index-html
+
+
+
+
+
 ## [1.9.2](https://github.com/open-wc/open-wc/compare/rollup-plugin-index-html@1.9.1...rollup-plugin-index-html@1.9.2) (2020-02-10)
 
 **Note:** Version bump only for package rollup-plugin-index-html